backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "celery>=5.6.0",
    "fastapi>=0.123.4",
    "httpx>=0.28.1",
    "huggingface-hub>=1.1.7",
    "lxml>=6.0.2",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "orjson>=3.11.4",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.44",
    "twilio>=9.8.7",
    "uvicorn[standard]>=0.38.0",
    "aiosqlite>=0.21.0",
    "bcrypt>=5.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "ruff>=0.14.7",
]
